move layout components to PageLayout folder
