--- src/components/Footer/index.tsx
@@ -1,1 +1,0 @@
-export * from './Footer'
