--- src/components/MenuMain/index.tsx
@@ -1,1 +1,0 @@
-export * from './MenuMain'
