+++ src/components/PageLayout/components/MenuSeminars.tsx
