+++ src/pages/Router/Router.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {Route, Switch, useRouteMatch} from 'react-router-dom'
 
-import {PageLayout} from '../../components/Layout/PageLayout'
+import {PageLayout} from '../../components/PageLayout/PageLayout'
 import {Admin} from '../Admin/Admin'
 import {PagePlaceholder} from '../PagePlaceholder'
